Change the types of the arguments from Integer/Float classes to primitive double

+++ src/main/java/dot503/Calculator.java
@@ -8,27 +8,27 @@ import org.springframework.web.bind.annotation.RestController;
 public class Calculator {
 
     @GetMapping("/addition/{a}/{b}")
-    public String Addition(@PathVariable Integer a, @PathVariable Integer b) {
+    public String Addition(@PathVariable double a, @PathVariable double b) {
         return String.valueOf(a + b);
     }
 
     @GetMapping("/substraction/{a}/{b}")
-    public String Substraction(@PathVariable Integer a, @PathVariable Integer b) {
+    public String Substraction(@PathVariable double a, @PathVariable double b) {
         return String.valueOf(a - b);
     }
 
     @GetMapping("/multiplication/{a}/{b}")
-    public String Multiplication(@PathVariable Integer a, @PathVariable Integer b) {
+    public String Multiplication(@PathVariable double a, @PathVariable double b) {
         return String.valueOf(a * b);
     }
 
     @GetMapping("/division/{a}/{b}")
-    public String Division(@PathVariable Float a, @PathVariable Float b) {
+    public String Division(@PathVariable double a, @PathVariable double b) {
         return String.valueOf(a / b);
     }
 
     @GetMapping("/power/{a}/{b}")
-    public String Power(@PathVariable Integer a, @PathVariable Integer b) {
-        return String.valueOf(Math.pow(Double.valueOf(a), Double.valueOf(b)));
+    public String Power(@PathVariable double a, @PathVariable double b) {
+        return String.valueOf(Math.pow(a, b));
     }
 };
